Create dedicate initializer for JavaRosa
